Add recipe picture to form values state
