+++ src/components/RecipeImage/RecipeImage.js
@@ -1,16 +1,20 @@
 import { useState } from 'react'
 import NextImage from 'next/image'
 import defaultPicture from 'public/images/default-recipe-picture.jpg'
-import { toBase64 } from '@/utils/toBase64'
-import { resizeImage } from '@/utils/resizeImage'
+import { toBase64, resizeImage } from '@/helpers/images'
 import styles from './RecipeImage.module.css'
 
-export default function RecipeImage() {
+export default function RecipeImage(props) {
   const [picture, setPicture] = useState()
 
   const onChangePicture = async (e) => {
     const file = await toBase64(e.target.files[0])
-    setPicture(resizeImage(file, 130, 130))
+    const newPicture = resizeImage(file, 130, 130)
+    setPicture(newPicture)
+    props.onChange(prev => ({
+      ...prev,
+      picture
+    }))
   }
 
   return (
