--- src/utils/toBase64.js
@@ -1,8 +1,0 @@
-export const toBase64 = (file) => {
-	return new Promise((resolve, reject) => {
-		const reader = new FileReader()
-		reader.readAsDataURL(file)
-		reader.onload = () => resolve(reader.result)
-		reader.onerror = (error) => reject(error)
-	})
-}
